auto-snapshot: before claude edit

+++ models/training/sft.py
@@ -39,7 +39,7 @@ def run_sft():
         "learning_rate": 2e-4,
         "lr_scheduler_type": "cosine",
         "warmup_ratio": 0.1,
-        "bf16": True,
+        "fp16": True,
         "save_steps": 50,
         "logging_steps": 10
     }
